+++ @navikt/core/react/src/expansion-card/ExpansionCardTitle.tsx
@@ -1,5 +1,5 @@
-import React, { forwardRef } from "react";
 import cl from "clsx";
+import React, { forwardRef } from "react";
 import { OverridableComponent } from "../util";
 
 export interface ExpansionCardTitleProps
